+++ _examples/3_DataCleaner.py
@@ -10,7 +10,7 @@ inv = read_inventory('data/XS.S11_decimated.station.xml', 'STATIONXML')
 dc = DataCleaner(stream, ['*1','*2','*H'])
 
 # Clean the data, then construct a stream with original and cleaned channels
-stream_cleaned = dc.clean_stream(stream)
+stream_cleaned = dc.apply(stream)
 z_compare = stream.select(channel='*Z') + stream_cleaned.select(channel='*Z')
 
 # If you print and plot the stream "normally", both channels have the same name