+++ tests/test_clean_rotator.py
@@ -30,7 +30,7 @@ class TestMethods(unittest.TestCase):
         stream = stream_read(str(self.test_path / 'XS.S10D.LH.mseed'))
         rotator = CleanRotator(
             stream, verbose=False,
-            remove_eq=str(self.test_path /
+            remove_eqs=str(self.test_path /
                           "20161205T-20161207T_MM5.85_eqcat.qml"),
             save_eq_file=False)
         self.assertAlmostEqual(rotator.angle, 0.18, delta=0.01)